fixes/improvements made at work: - fix lots of 64-bit warnings - round_up/down now templates - avoid warning with definition of byte swap functions; remove duplication of that in wsdl - codec_zlib.cpp: avoid domination warning - vfs/file_cache: VFS is now responsible for handling zero-length files (no longer considered an error; just returns zero pointer and size=0) - cpu: export all functions (thus obviating cpu_memcpy_thunk). this required renaming asm functions and adding thunk functions that call them - winit: fix segment merge statement, reinstate /include (otherwise init functions are stripped by linker) - wstartup: VC CRT init section definitions have changed in amd64 build; match their definition - wsysdep.cpp: more descriptive text for osError = 0
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@5899 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/allocators/headerless.cpp
@@ -50,7 +50,7 @@ public:
 		// clear all fields to prevent accidental reuse
 		prev = next = 0;
 		m_id = 0;
-		m_size = ~0u;
+		m_size = ~(size_t)0u;
 		m_magic = 0;
 	}
